removed local preprocessing from torch training test and unpinned h5py
