+++ test/sagemaker_tests/huggingface_tensorflow/training/requirements.txt
@@ -10,7 +10,7 @@ scipy
 sklearn
 pandas
 Pillow
-h5py==2.10.0
+h5py
 tox
 flake8
 pytest==5.3.5
@@ -30,6 +30,4 @@ fabric==2.5.0
 retrying==1.3.3
 gitpython
 toml
-s3fs>=0.4.2
-datasets[s3]==1.6.2
-transformers==4.6.1
+
